Películas controller terminado (en teoría)
Ya se pueden hacer todas las operaciones relacionadas con las películas

+++ contenidosflask/openapi_server/controllers/pelicula_controller.py
@@ -1,4 +1,5 @@
 from openapi_server.models.actor import Actor  # noqa: E501
+from openapi_server.models.actor_db import ActorDB  # noqa: E501
 from openapi_server.models.asignar_actor_a_serie_request import AsignarActorASerieRequest  # noqa: E501
 from openapi_server.models.pelicula import Pelicula  # noqa: E501
 from openapi_server.models.pelicula_update import PeliculaUpdate  # noqa: E501
@@ -10,8 +11,8 @@ from flask import jsonify
 from bson import ObjectId
 from openapi_server.models.pelicula_db import PeliculaDB
 
-
-def actualizar_pelicula(pelicula_id, pelicula_update):  # noqa: E501
+@app.route('/actualizar_pelicula/<pelicula_id>', methods=['PUT'])
+def actualizar_pelicula(pelicula_id):  # noqa: E501
     """Actualizar una película existente
 
     Actualiza la información de una película específica por su ID. # noqa: E501
@@ -25,10 +26,27 @@ def actualizar_pelicula(pelicula_id, pelicula_update):  # noqa: E501
     """
     if request.is_json:
         pelicula_update = PeliculaUpdate.from_dict(request.get_json())  # noqa: E501
-    return 'do some magic!'
 
+    pelicula_to_update = PeliculaDB.objects.get(id=ObjectId(pelicula_id))
+    if not pelicula_to_update:
+        return jsonify({"message": "Película no encontrada", "status": "error"}), 404
+    
+    pelicula_db = pelicula_update.to_db_model()
+    pelicula_to_update.titulo = pelicula_db.titulo
+    pelicula_to_update.genero = pelicula_db.genero
+    pelicula_to_update.sinopsis = pelicula_db.sinopsis
+    pelicula_to_update.anio_estreno = pelicula_db.anio_estreno
+    pelicula_to_update.duracion = pelicula_db.duracion
 
-def asignar_actor_a_pelicula(pelicula_id, asignar_actor_a_serie_request):  # noqa: E501
+    # Reemplazar los actores
+    actores_db = [ActorDB.objects.get(id=ObjectId(id)) for id in pelicula_update.actores]
+    pelicula_to_update.actores = actores_db # Se cambia la lista de actores por la nueva
+
+    pelicula_to_update.save()
+    return jsonify({"message": "Película actualizada correctamente", "status": "success"}), 200
+
+@app.route('/asignar_actor_pelicula/<pelicula_id>', methods=['POST'])
+def asignar_actor_a_pelicula(pelicula_id):  # noqa: E501
     """Asignar un actor a una película
 
     Asigna un actor existente a una película específica. # noqa: E501
@@ -40,9 +58,19 @@ def asignar_actor_a_pelicula(pelicula_id, asignar_actor_a_serie_request):  # noq
 
     :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
     """
+    
     if request.is_json:
-        asignar_actor_a_serie_request = AsignarActorASerieRequest.from_dict(request.get_json())  # noqa: E501
-    return 'do some magic!'
+        actor_a_asignar = AsignarActorASerieRequest.from_dict(request.get_json())
+
+    pelicula_db = PeliculaDB.objects.get(id=ObjectId(pelicula_id))
+    if not pelicula_db:
+        return jsonify({"message": "Película no encontrada", "status": "error"}), 404
+
+    if not ActorDB.objects.get(id=ObjectId(actor_a_asignar.actor_id)):
+        return jsonify({"message": "Actor no encontrado", "status": "error"}), 404
+
+    pelicula_db.update(add_to_set__actores=actor_a_asignar.actor_id)
+    return jsonify({"message": "Actor asignado con éxito", "status": "success"}), 200
 
 @app.route('/crear_pelicula', methods=['POST'])
 def crear_pelicula():  # noqa: E501
@@ -57,9 +85,13 @@ def crear_pelicula():  # noqa: E501
     """
     if request.is_json:
         pelicula_api = Pelicula.from_dict(request.get_json())
+
+    if (pelicula_api):
         pelicula_db = pelicula_api.to_db_model()
         pelicula_db.save()
         return jsonify({"message": "Película creada con éxito", "status": "success"}), 201
+    else:
+        return jsonify({"message": "Error al crear la película", "status": "error"}), 400
 
 
 @app.route('/eliminar_actor_pelicula/<pelicula_id>/<actor_id>', methods=['DELETE'])
@@ -76,8 +108,14 @@ def eliminar_actor_pelicula(pelicula_id, actor_id):  # noqa: E501
     :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
     """
     pelicula_db = PeliculaDB.objects.get(id=ObjectId(pelicula_id))
-    pelicula_db.actores.remove(ObjectId(actor_id))
-    pelicula_db.save()
+    if not pelicula_db:
+        return jsonify({"message": "Película no encontrada", "status": "error"}), 404 
+
+    actor_db = ActorDB.objects.get(id=ObjectId(actor_id))
+    if not actor_db in pelicula_db.actores:
+        return jsonify({"message": "Actor no encontrado en la película", "status": "error"}), 404
+
+    pelicula_db.update(pull__actores=actor_db)
     return jsonify({"message": "Actor eliminado con éxito", "status": "success"}), 200
 
 
@@ -94,10 +132,13 @@ def eliminar_pelicula(pelicula_id):  # noqa: E501
     """
 
     pelicula_db = PeliculaDB.objects.get(id=ObjectId(pelicula_id))
+    if not pelicula_db:
+        return jsonify({"message": "Película no encontrada", "status": "error"}), 404
+
     pelicula_db.delete()
     return jsonify({"message": "Película eliminada con éxito", "status": "success"}), 200
 
-@app.route('/listar_actores_de_pelicula/<pelicula_id>', methods=['GET'])
+@app.route('/listar_actores_pelicula/<pelicula_id>', methods=['GET'])
 def listar_actores_de_pelicula(pelicula_id):  # noqa: E501
     """Listar actores de una película específica
 
@@ -110,9 +151,12 @@ def listar_actores_de_pelicula(pelicula_id):  # noqa: E501
     """
 
     pelicula_db = PeliculaDB.objects.get(id=ObjectId(pelicula_id))
+    if not pelicula_db:
+        return jsonify({"message": "Película no encontrada", "status": "error"}), 404
+
     pelicula_api = pelicula_db.to_api_model()
-    actores_json = [actor.serialize() for actor in pelicula_api.actores]
-    return jsonify(actores_json)
+    actores_json = [actor for actor in pelicula_api.actores]
+    return jsonify(actores_json), 200
     
 
 @app.route('/listar_peliculas', methods=['GET'])
@@ -125,9 +169,10 @@ def listar_peliculas():  # noqa: E501
     :rtype: Union[List[Pelicula], Tuple[List[Pelicula], int], Tuple[List[Pelicula], int, Dict[str, str]]
     """
     peliculas_db = PeliculaDB.objects()
-    return [pelicula.to_api_model() for pelicula in peliculas_db]
+    list_peliculas = [pelicula.to_api_model() for pelicula in peliculas_db]
+    return jsonify(list_peliculas), 200
 
-@app.route('/pelicula/<pelicula_id>', methods=['GET'])
+@app.route('/obtener_pelicula/<pelicula_id>', methods=['GET'])
 def obtener_pelicula(pelicula_id):  # noqa: E501
     """Obtener una película específica
 
@@ -139,10 +184,11 @@ def obtener_pelicula(pelicula_id):  # noqa: E501
     :rtype: Union[Pelicula, Tuple[Pelicula, int], Tuple[Pelicula, int, Dict[str, str]]
     """
     pelicula_db = PeliculaDB.objects.get(id=ObjectId(pelicula_id))
-    return jsonify(pelicula_db.to_api_model())
+    if not pelicula_db:
+        return jsonify({"message": "Película no encontrada", "status": "error"}), 404
     
-
-
+    return jsonify(pelicula_db.to_api_model()), 200
+    
 def obtener_precuela_pelicula(pelicula_id):  # noqa: E501
     """Obtiene la precuela de una película específica
